requirements.txt
blinker
click
dnspython
Flask
itsdangerous
Jinja2
MarkupSafe
pymongo
Werkzeug
gunicorn
flask-wtf
python-dotenv
xmltodict
email_validator